Added header file and main cpp
