+++ ClassStudent.h
@@ -1,0 +1,22 @@
+#ifndef ClassStudent_h
+#define ClassStudent_h
+
+using namespace std;
+
+class Student
+{
+public:
+    
+    string name;
+    string surname;
+    string secondName;
+    string dateOfBirth;
+    string id;
+    string departmentId;
+    string diaryId;
+    string learningProgrammingId;
+    string courseStartDate;
+    
+};
+
+#endif /* ClassStudent_h */
